+++ mocks/fix_tool.py
@@ -287,6 +287,9 @@ def delete_old_order():
     utc_now = datetime.datetime.now().astimezone(pytz.utc)
     orders = Orders.objects.filter(expired_at__lt=utc_now - datetime.timedelta(days=30)).all()
     for order_obj_item in orders:
+        oerder_id = order_obj_item.id
+        if Proxy.objects.filter(order_id=oerder_id).exists():
+            continue
         delete_list.append(order_obj_item.id)
         order_obj_item.delete()
     data = {
